+++ src/main/Descuentos/Desc2Meses.java
@@ -3,7 +3,7 @@ package main.Descuentos;
 public class Desc2Meses extends Descuento {
 
     //constructor
-    public Desc2Meses(float descuento) {
+    public Desc2Meses(Double descuento) {
         super(descuento);
     }
