+++ src/main/Descuentos/SinDesc.java
@@ -3,7 +3,7 @@ package main.Descuentos;
 public class SinDesc extends Descuento{
 
     //constructor
-    public SinDesc(float descuento) {
+    public SinDesc(Double descuento) {
         super(descuento);
     }
